+++ Exercises/Part-8/GraphQL-backend/index.js
@@ -3,8 +3,6 @@ const { startStandaloneServer } = require('@apollo/server/standalone')
 const mongoose = require('mongoose')
 const { GraphQLError } = require('graphql')
 mongoose.set('strictQuery', false)
-const Author = require('./models/Author')
-const Book = require('./models/Book')
 const User = require('./models/User')
 const jwt = require('jsonwebtoken')
 
@@ -22,345 +20,6 @@ mongoose
   .catch((error) => {
     console.log('error connecting to MongoDB:', error.message)
   })
-
-let authors = [
-  {
-    name: 'Robert Martin',
-    id: 'afa51ab0-344d-11e9-a414-719c6709cf3e',
-    born: 1952,
-  },
-  {
-    name: 'Martin Fowler',
-    id: 'afa5b6f0-344d-11e9-a414-719c6709cf3e',
-    born: 1963,
-  },
-  {
-    name: 'Fyodor Dostoevsky',
-    id: 'afa5b6f1-344d-11e9-a414-719c6709cf3e',
-    born: 1821,
-  },
-  {
-    name: 'Joshua Kerievsky', // birthyear not known
-    id: 'afa5b6f2-344d-11e9-a414-719c6709cf3e',
-  },
-  {
-    name: 'Sandi Metz', // birthyear not known
-    id: 'afa5b6f3-344d-11e9-a414-719c6709cf3e',
-  },
-]
-
-/*
- * Suomi:
- * Saattaisi olla järkevämpää assosioida kirja ja sen tekijä tallettamalla kirjan yhteyteen tekijän nimen sijaan tekijän id
- * Yksinkertaisuuden vuoksi tallennamme kuitenkin kirjan yhteyteen tekijän nimen
- *
- * English:
- * It might make more sense to associate a book with its author by storing the author's id in the context of the book instead of the author's name
- * However, for simplicity, we will store the author's name in connection with the book
- *
- * Spanish:
- * Podría tener más sentido asociar un libro con su autor almacenando la id del autor en el contexto del libro en lugar del nombre del autor
- * Sin embargo, por simplicidad, almacenaremos el nombre del autor en conexión con el libro
- */
-
-let books = [
-  {
-    title: 'Clean Code',
-    published: 2008,
-    author: 'Robert Martin',
-    id: 'afa5b6f4-344d-11e9-a414-719c6709cf3e',
-    genres: ['refactoring'],
-  },
-  {
-    title: 'Agile software development',
-    published: 2002,
-    author: 'Robert Martin',
-    id: 'afa5b6f5-344d-11e9-a414-719c6709cf3e',
-    genres: ['agile', 'patterns', 'design'],
-  },
-  {
-    title: 'Refactoring, edition 2',
-    published: 2018,
-    author: 'Martin Fowler',
-    id: 'afa5de00-344d-11e9-a414-719c6709cf3e',
-    genres: ['refactoring'],
-  },
-  {
-    title: 'Refactoring to patterns',
-    published: 2008,
-    author: 'Joshua Kerievsky',
-    id: 'afa5de01-344d-11e9-a414-719c6709cf3e',
-    genres: ['refactoring', 'patterns'],
-  },
-  {
-    title: 'Practical Object-Oriented Design, An Agile Primer Using Ruby',
-    published: 2012,
-    author: 'Sandi Metz',
-    id: 'afa5de02-344d-11e9-a414-719c6709cf3e',
-    genres: ['refactoring', 'design'],
-  },
-  {
-    title: 'Crime and punishment',
-    published: 1866,
-    author: 'Fyodor Dostoevsky',
-    id: 'afa5de03-344d-11e9-a414-719c6709cf3e',
-    genres: ['classic', 'crime'],
-  },
-  {
-    title: 'Demons',
-    published: 1872,
-    author: 'Fyodor Dostoevsky',
-    id: 'afa5de04-344d-11e9-a414-719c6709cf3e',
-    genres: ['classic', 'revolution'],
-  },
-]
-
-/*const populateDB = async () => {
-  await Book.deleteMany({})
-  await Author.deleteMany({})
-
-  //await Book.insertMany(books)
-  await Author.insertMany(authors)
-}
-
-populateDB()*/
-
-const typeDefs = `
-
-  type Author {
-    name: String!
-    id: ID!
-    born: Int
-    bookCount: Int
-  }
-
-  type Book {
-    title: String!
-    published: Int
-    author: Author!
-    id: ID!
-    genres: [String!]!
-  }
-
-  type User {
-    username: String!
-    favoriteGenre: String!
-    id: ID!
-  }
-
-  type Token {
-    value: String!
-  }
-
-  type Query {
-    bookCount: Int!
-    authorCount: Int!
-    allBooks(author: String, genre: String): [Book!]!
-    allAuthors: [Author!]!
-    me: User
-  }
-
-  type Mutation {
-    addBook(
-      title: String!
-      author: String!
-      published: Int
-      genres: [String!]!
-    ): Book
-
-    editAuthor(
-      name: String!
-      setBornTo: Int!
-    ): Author
-
-    createUser(
-      username: String!
-      favoriteGenre: String!  
-    ): User
-
-    login(
-      username: String!
-      password: String!
-    ): Token
-  }
-`
-
-const resolvers = {
-  Query: {
-    bookCount: async () => Book.collection.countDocuments(),
-    authorCount: async () => Author.collection.countDocuments(),
-    allBooks: async (root, args) => {
-      //can be optimized but left for readibility
-      if (args.author && args.genre) {
-        const author = await Author.findOne({ name: args.author })
-        if (!author) {
-          throw new GraphQLError('author not found')
-        }
-        const authorId = author._id.toString()
-        const book = await Book.find({
-          genres: args.genre,
-          author: authorId,
-        }).populate('author')
-        console.log(book)
-        return book
-      } else if (args.author) {
-        const author = await Author.findOne({ name: args.author })
-        if (!author) {
-          throw new GraphQLError('author not found')
-        }
-
-        const authorId = author._id.toString()
-        const book = await Book.find({ author: authorId }).populate('author')
-
-        return book
-      } else if (args.genre) {
-        return Book.find({ genres: args.genre }).populate('author')
-      } else {
-        return Book.find({}).populate('author')
-      }
-    },
-    allAuthors: async () => {
-      return Author.find({})
-    },
-    me: (root, args, context) => {
-      return context.currentUser
-    },
-  },
-  Author: {
-    bookCount: async (root) => {
-      const author = await Author.findOne({ name: root.name })
-      const authorId = author._id.toString()
-      const booksDB = await Book.find({})
-      let count = 0
-      for (let i of booksDB) {
-        if (i.author.toString() === authorId) {
-          count++
-        }
-      }
-      return count
-    },
-  },
-  Mutation: {
-    addBook: async (root, args, context) => {
-      let authorId = ''
-
-      const currentUser = context.currentUser
-      if (!currentUser) {
-        throw new GraphQLError('not authenticated', {
-          extensions: {
-            code: 'BAD_USER_INPUT',
-          },
-        })
-      }
-
-      let existingAuthor = await Author.findOne({ name: args.author })
-      if (!existingAuthor) {
-        const author = new Author({ name: args.author })
-
-        try {
-          await author.save()
-          console.log(author)
-          authorId = author._id.toString()
-        } catch (error) {
-          throw new GraphQLError('saving author failed', {
-            extensions: {
-              code: 'BAD_USER_INPUT',
-              invalidArgs: args.name,
-              error,
-            },
-          })
-        }
-      } else {
-        authorId = existingAuthor._id.toString()
-      }
-      const book = new Book({ ...args, author: authorId })
-      console.log(book)
-      try {
-        await book.save()
-      } catch (error) {
-        console.log(error)
-        throw new GraphQLError('saving book failed', {
-          extensions: {
-            code: 'BAD_USER_INPUT',
-            invalidArgs: args.name,
-            error,
-          },
-        })
-      }
-
-      return book.populate('author')
-    },
-    editAuthor: async (root, args, context) => {
-      const currentUser = context.currentUser
-      if (!currentUser) {
-        throw new GraphQLError('not authenticated', {
-          extensions: {
-            code: 'BAD_USER_INPUT',
-          },
-        })
-      }
-
-      const author = await Author.findOne({ name: args.name })
-      if (!author) {
-        return null
-      }
-
-      author.born = args.setBornTo
-
-      try {
-        await author.save()
-      } catch (error) {
-        console.log(error)
-        throw new GraphQLError('Editing failed', {
-          extensions: {
-            code: 'BAD_USER_INPUT',
-            invalidArgs: args.name,
-            error,
-          },
-        })
-      }
-
-      return author
-    },
-    createUser: async (root, args) => {
-      const user = new User({
-        username: args.username,
-        favoriteGenre: args.favoriteGenre,
-      })
-      console.log(user)
-
-      return user.save().catch((error) => {
-        throw new GraphQLError('creating user failed', {
-          extensions: {
-            code: 'BAD_USER_INPUT',
-            invalidArgs: args.name,
-            error,
-          },
-        })
-      })
-    },
-    login: async (root, args) => {
-      const user = await User.findOne({ username: args.username })
-
-      if (!user || args.password !== 'secret') {
-        throw new GraphQLError('wrong credentials', {
-          extensions: {
-            code: 'BAD_USER_INPUT',
-            invalidArgs: args.name,
-            error,
-          },
-        })
-      }
-
-      const userForToken = {
-        username: user.username,
-        id: user._id,
-      }
-
-      return { value: jwt.sign(userForToken, process.env.JWT_SECRET) }
-    },
-  },
-}
 
 const server = new ApolloServer({
   typeDefs,
