+++ Exercises/Part-8/GraphQL-backend/resolvers.js
@@ -1,0 +1,184 @@
+const { GraphQLError } = require('graphql')
+const jwt = require('jsonwebtoken')
+const Author = require('./models/Author')
+const Book = require('./models/Book')
+const User = require('./models/User')
+
+const resolvers = {
+  Query: {
+    bookCount: async () => Book.collection.countDocuments(),
+    authorCount: async () => Author.collection.countDocuments(),
+    allBooks: async (root, args) => {
+      //can be optimized but left for readibility
+      if (args.author && args.genre) {
+        const author = await Author.findOne({ name: args.author })
+        if (!author) {
+          throw new GraphQLError('author not found')
+        }
+        const authorId = author._id.toString()
+        const book = await Book.find({
+          genres: args.genre,
+          author: authorId,
+        }).populate('author')
+        console.log(book)
+        return book
+      } else if (args.author) {
+        const author = await Author.findOne({ name: args.author })
+        if (!author) {
+          throw new GraphQLError('author not found')
+        }
+
+        const authorId = author._id.toString()
+        const book = await Book.find({ author: authorId }).populate('author')
+
+        return book
+      } else if (args.genre) {
+        return Book.find({ genres: args.genre }).populate('author')
+      } else {
+        return Book.find({}).populate('author')
+      }
+    },
+    allAuthors: async () => {
+      return Author.find({})
+    },
+    me: (root, args, context) => {
+      return context.currentUser
+    },
+  },
+  Author: {
+    bookCount: async (root) => {
+      const author = await Author.findOne({ name: root.name })
+      const authorId = author._id.toString()
+      const booksDB = await Book.find({})
+      let count = 0
+      for (let i of booksDB) {
+        if (i.author.toString() === authorId) {
+          count++
+        }
+      }
+      return count
+    },
+  },
+  Mutation: {
+    addBook: async (root, args, context) => {
+      let authorId = ''
+
+      const currentUser = context.currentUser
+      if (!currentUser) {
+        throw new GraphQLError('not authenticated', {
+          extensions: {
+            code: 'BAD_USER_INPUT',
+          },
+        })
+      }
+
+      let existingAuthor = await Author.findOne({ name: args.author })
+      if (!existingAuthor) {
+        const author = new Author({ name: args.author })
+
+        try {
+          await author.save()
+          console.log(author)
+          authorId = author._id.toString()
+        } catch (error) {
+          throw new GraphQLError('saving author failed', {
+            extensions: {
+              code: 'BAD_USER_INPUT',
+              invalidArgs: args.name,
+              error,
+            },
+          })
+        }
+      } else {
+        authorId = existingAuthor._id.toString()
+      }
+      const book = new Book({ ...args, author: authorId })
+      console.log(book)
+      try {
+        await book.save()
+      } catch (error) {
+        console.log(error)
+        throw new GraphQLError('saving book failed', {
+          extensions: {
+            code: 'BAD_USER_INPUT',
+            invalidArgs: args.name,
+            error,
+          },
+        })
+      }
+
+      return book.populate('author')
+    },
+    editAuthor: async (root, args, context) => {
+      const currentUser = context.currentUser
+      if (!currentUser) {
+        throw new GraphQLError('not authenticated', {
+          extensions: {
+            code: 'BAD_USER_INPUT',
+          },
+        })
+      }
+
+      const author = await Author.findOne({ name: args.name })
+      if (!author) {
+        return null
+      }
+
+      author.born = args.setBornTo
+
+      try {
+        await author.save()
+      } catch (error) {
+        console.log(error)
+        throw new GraphQLError('Editing failed', {
+          extensions: {
+            code: 'BAD_USER_INPUT',
+            invalidArgs: args.name,
+            error,
+          },
+        })
+      }
+
+      return author
+    },
+    createUser: async (root, args) => {
+      const user = new User({
+        username: args.username,
+        favoriteGenre: args.favoriteGenre,
+      })
+      console.log(user)
+
+      return user.save().catch((error) => {
+        throw new GraphQLError('creating user failed', {
+          extensions: {
+            code: 'BAD_USER_INPUT',
+            invalidArgs: args.name,
+            error,
+          },
+        })
+      })
+    },
+    login: async (root, args) => {
+      const user = await User.findOne({ username: args.username })
+
+      if (!user || args.password !== 'secret') {
+        throw new GraphQLError('wrong credentials', {
+          extensions: {
+            code: 'BAD_USER_INPUT',
+            invalidArgs: args.name,
+            error,
+          },
+        })
+      }
+
+      const userForToken = {
+        username: user.username,
+        id: user._id,
+      }
+
+      return { value: jwt.sign(userForToken, process.env.JWT_SECRET) }
+    },
+  },
+}
+
+module.exports = resolvers
